move query utils to new folder
